ini/lucIA/requirements.txt
# Dependencias para LucIA - IA de la Plataforma Metaverso
# APIs y comunicación
aiohttp==3.9.1
requests==2.31.0
openai==1.3.7
anthropic==0.7.8
google-generativeai==0.3.2

# Base de datos y almacenamiento
sqlite3
pickle-mixin==1.0.2

# Procesamiento de texto y NLP
nltk==3.8.1
textblob==0.17.1
spacy==3.7.2
transformers==4.36.2
torch==2.1.2

# Utilidades
python-dotenv==1.0.0
asyncio
dataclasses
typing-extensions==4.8.0
pydantic==2.5.2

# Logging y monitoreo
loguru==0.7.2
rich==13.7.0

# Criptografía para seguridad
cryptography==41.0.8
hashlib

# Cache y optimización
redis==5.0.1
cachetools==5.3.2

# Testing
pytest==7.4.4
pytest-asyncio==0.21.1

# Desarrollo
black==23.12.1
flake8==6.1.0
mypy==1.8.0 